requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
PyYAML==6.0.2
beautifulsoup4==4.12.3
